Algorithm for finding maximum satisfaction for eating at a restaurant
A restaurant has different types of food to choose from. A person gets
x amount of satisfaction and requires y amount of time to eat an item
from the menu. Given t minutes, the algorithm ﬁnds out the maximum
satisfaction that the person can get from eating at the restaurant.